// src/app/models/transaction.ts
export interface Transaction {
    lpoCode: string;
    lpoName: string;
    projectCode: string;
    transactionType: string;
    transactionDescription: string;
    transactionDate: string;
    amount: number;
    quantity: number;
    unitDescription: string;
    unitPrice: number;
    deliveryNoteNumber: string;
    vehicleNumber: string;
}